+++ prajna/transform/verify.hpp
@@ -42,7 +42,7 @@ inline void VerifyBlockImpl(std::shared_ptr<ir::Block> ir_block,
         }
 
         if (auto ir_instruction = Cast<ir::Instruction>(ir_value)) {
-            for (size_t i = 0; i < ir_instruction->OperandSize(); ++i) {
+            for (int64_t i = 0; i < ir_instruction->OperandSize(); ++i) {
                 auto ir_operand = ir_instruction->GetOperand(i);
                 PRAJNA_ASSERT(ir_operand);
 
@@ -82,7 +82,7 @@ inline bool VerifyModule(std::shared_ptr<ir::Module> ir_module) {
                 continue;
             }
 
-            for (size_t i = 0; i < ir_instruction->OperandSize(); ++i) {
+            for (int64_t i = 0; i < ir_instruction->OperandSize(); ++i) {
                 auto ir_operand = ir_instruction->GetOperand(i);
 
                 PRAJNA_ASSERT(ir_operand);
